Refactor common functions in planner and smoother
