+++ nav2_smac_planner/src/smoother.cpp
@@ -24,11 +24,13 @@
 #include "tf2/utils.hpp"
 
 #include "nav2_smac_planner/smoother.hpp"
+#include "nav2_util/smoother_utils.hpp"
 
 namespace nav2_smac_planner
 {
 using namespace nav2_util::geometry_utils;  // NOLINT
 using namespace std::chrono;  // NOLINT
+using nav2_util::PathSegment;
 
 Smoother::Smoother(const SmootherParams & params)
 {
@@ -62,7 +64,7 @@ bool Smoother::smooth(
   bool success = true, reversing_segment;
   nav_msgs::msg::Path curr_path_segment;
   curr_path_segment.header = path.header;
-  std::vector<PathSegment> path_segments = findDirectionalPathSegments(path);
+  std::vector<PathSegment> path_segments = nav2_util::findDirectionalPathSegments(path);
 
   for (unsigned int i = 0; i != path_segments.size(); i++) {
     if (path_segments[i].end - path_segments[i].start > 10) {
@@ -130,7 +132,7 @@ bool Smoother::smoothImpl(
         rclcpp::get_logger("SmacPlannerSmoother"),
         "Number of iterations has exceeded limit of %i.", max_its_);
       path = last_path;
-      updateApproximatePathOrientations(path, reversing_segment);
+      nav2_util::updateApproximatePathOrientations(path, reversing_segment);
       return false;
     }
 
@@ -142,7 +144,7 @@ bool Smoother::smoothImpl(
         rclcpp::get_logger("SmacPlannerSmoother"),
         "Smoothing time exceeded allowed duration of %0.2f.", max_time);
       path = last_path;
-      updateApproximatePathOrientations(path, reversing_segment);
+      nav2_util::updateApproximatePathOrientations(path, reversing_segment);
       return false;
     }
 
@@ -176,7 +178,7 @@ bool Smoother::smoothImpl(
           "Smoothing process resulted in an infeasible collision. "
           "Returning the last path before the infeasibility was introduced.");
         path = last_path;
-        updateApproximatePathOrientations(path, reversing_segment);
+        nav2_util::updateApproximatePathOrientations(path, reversing_segment);
         return false;
       }
     }
@@ -191,7 +193,7 @@ bool Smoother::smoothImpl(
     smoothImpl(new_path, reversing_segment, costmap, max_time);
   }
 
-  updateApproximatePathOrientations(new_path, reversing_segment);
+  nav2_util::updateApproximatePathOrientations(new_path, reversing_segment);
   path = new_path;
   return true;
 }
@@ -218,92 +220,6 @@ void Smoother::setFieldByDim(
     msg.pose.position.y = value;
   } else {
     msg.pose.position.z = value;
-  }
-}
-
-std::vector<PathSegment> Smoother::findDirectionalPathSegments(const nav_msgs::msg::Path & path)
-{
-  std::vector<PathSegment> segments;
-  PathSegment curr_segment;
-  curr_segment.start = 0;
-
-  // If holonomic, no directional changes and
-  // may have abrupt angular changes from naive grid search
-  if (is_holonomic_) {
-    curr_segment.end = path.poses.size() - 1;
-    segments.push_back(curr_segment);
-    return segments;
-  }
-
-  // Iterating through the path to determine the position of the cusp
-  for (unsigned int idx = 1; idx < path.poses.size() - 1; ++idx) {
-    // We have two vectors for the dot product OA and AB. Determining the vectors.
-    double oa_x = path.poses[idx].pose.position.x -
-      path.poses[idx - 1].pose.position.x;
-    double oa_y = path.poses[idx].pose.position.y -
-      path.poses[idx - 1].pose.position.y;
-    double ab_x = path.poses[idx + 1].pose.position.x -
-      path.poses[idx].pose.position.x;
-    double ab_y = path.poses[idx + 1].pose.position.y -
-      path.poses[idx].pose.position.y;
-
-    // Checking for the existence of cusp, in the path, using the dot product.
-    double dot_product = (oa_x * ab_x) + (oa_y * ab_y);
-    if (dot_product < 0.0) {
-      curr_segment.end = idx;
-      segments.push_back(curr_segment);
-      curr_segment.start = idx;
-    }
-
-    // Checking for the existence of a differential rotation in place.
-    double cur_theta = tf2::getYaw(path.poses[idx].pose.orientation);
-    double next_theta = tf2::getYaw(path.poses[idx + 1].pose.orientation);
-    double dtheta = angles::shortest_angular_distance(cur_theta, next_theta);
-    if (fabs(ab_x) < 1e-4 && fabs(ab_y) < 1e-4 && fabs(dtheta) > 1e-4) {
-      curr_segment.end = idx;
-      segments.push_back(curr_segment);
-      curr_segment.start = idx;
-    }
-  }
-
-  curr_segment.end = path.poses.size() - 1;
-  segments.push_back(curr_segment);
-  return segments;
-}
-
-void Smoother::updateApproximatePathOrientations(
-  nav_msgs::msg::Path & path,
-  bool & reversing_segment)
-{
-  double dx, dy, theta, pt_yaw;
-  reversing_segment = false;
-
-  // Find if this path segment is in reverse
-  dx = path.poses[2].pose.position.x - path.poses[1].pose.position.x;
-  dy = path.poses[2].pose.position.y - path.poses[1].pose.position.y;
-  theta = atan2(dy, dx);
-  pt_yaw = tf2::getYaw(path.poses[1].pose.orientation);
-  if (!is_holonomic_ && fabs(angles::shortest_angular_distance(pt_yaw, theta)) > M_PI_2) {
-    reversing_segment = true;
-  }
-
-  // Find the angle relative the path position vectors
-  for (unsigned int i = 0; i != path.poses.size() - 1; i++) {
-    dx = path.poses[i + 1].pose.position.x - path.poses[i].pose.position.x;
-    dy = path.poses[i + 1].pose.position.y - path.poses[i].pose.position.y;
-    theta = atan2(dy, dx);
-
-    // If points are overlapping, pass
-    if (fabs(dx) < 1e-4 && fabs(dy) < 1e-4) {
-      continue;
-    }
-
-    // Flip the angle if this path segment is in reverse
-    if (reversing_segment) {
-      theta += M_PI;  // orientationAroundZAxis will normalize
-    }
-
-    path.poses[i].pose.orientation = orientationAroundZAxis(theta);
   }
 }
