+++ nav2_util/include/nav2_util/smoother_utils.hpp
@@ -12,31 +12,25 @@
 // See the License for the specific language governing permissions and
 // limitations under the License. Reserved.
 
-#ifndef NAV2_SMOOTHER__SMOOTHER_UTILS_HPP_
-#define NAV2_SMOOTHER__SMOOTHER_UTILS_HPP_
+#ifndef NAV2_UTIL__SMOOTHER_UTILS_HPP_
+#define NAV2_UTIL__SMOOTHER_UTILS_HPP_
 
 #include <cmath>
 #include <vector>
 #include <string>
-#include <iostream>
 #include <memory>
-#include <queue>
 #include <utility>
 
-#include "nav2_core/smoother.hpp"
-#include "nav2_costmap_2d/costmap_2d.hpp"
-#include "nav2_costmap_2d/cost_values.hpp"
 #include "nav2_util/geometry_utils.hpp"
-#include "nav2_ros_common/node_utils.hpp"
 #include "nav_msgs/msg/path.hpp"
 #include "angles/angles.h"
 #include "tf2/utils.hpp"
 
-namespace smoother_utils
+namespace nav2_util
 {
 
 /**
- * @class nav2_smoother::PathSegment
+ * @class nav2_util::PathSegment
  * @brief A segment of a path in start/end indices
  */
 struct PathSegment
@@ -45,15 +39,27 @@ struct PathSegment
   unsigned int end;
 };
 
-typedef std::vector<geometry_msgs::msg::PoseStamped>::iterator PathIterator;
-typedef std::vector<geometry_msgs::msg::PoseStamped>::reverse_iterator ReversePathIterator;
-
+/**
+ * @brief Finds the starting and end indices of path segments where
+ * the robot is traveling in the same direction (e.g. forward vs reverse)
+ * @param path Path in which to look for cusps
+ * @param is_holonomic Whether the motion model is holonomic (default is false)
+ * @return Set of index pairs for each segment of the path in a given direction
+ */
 inline std::vector<PathSegment> findDirectionalPathSegments(
-  const nav_msgs::msg::Path & path)
+  const nav_msgs::msg::Path & path, bool is_holonomic = false)
 {
   std::vector<PathSegment> segments;
   PathSegment curr_segment;
   curr_segment.start = 0;
+
+  // If holonomic, no directional changes and
+  // may have abrupt angular changes from naive grid search
+  if (is_holonomic) {
+    curr_segment.end = path.poses.size() - 1;
+    segments.push_back(curr_segment);
+    return segments;
+  }
 
   // Iterating through the path to determine the position of the cusp
   for (unsigned int idx = 1; idx < path.poses.size() - 1; ++idx) {
@@ -91,9 +97,16 @@ inline std::vector<PathSegment> findDirectionalPathSegments(
   return segments;
 }
 
+/**
+ * @brief For a given path, update the path point orientations based on smoothing
+ * @param path Path to approximate the path orientation in
+ * @param reversing_segment Return if this is a reversing segment
+ * @param is_holonomic Whether the motion model is holonomic (default is false)
+ */
 inline void updateApproximatePathOrientations(
   nav_msgs::msg::Path & path,
-  bool & reversing_segment)
+  bool & reversing_segment,
+  bool is_holonomic = false)
 {
   double dx, dy, theta, pt_yaw;
   reversing_segment = false;
@@ -103,7 +116,7 @@ inline void updateApproximatePathOrientations(
   dy = path.poses[2].pose.position.y - path.poses[1].pose.position.y;
   theta = atan2(dy, dx);
   pt_yaw = tf2::getYaw(path.poses[1].pose.orientation);
-  if (fabs(angles::shortest_angular_distance(pt_yaw, theta)) > M_PI_2) {
+  if (!is_holonomic && fabs(angles::shortest_angular_distance(pt_yaw, theta)) > M_PI_2) {
     reversing_segment = true;
   }
 
@@ -127,6 +140,6 @@ inline void updateApproximatePathOrientations(
   }
 }
 
-}  // namespace smoother_utils
+}  // namespace nav2_util
 
-#endif  // NAV2_SMOOTHER__SMOOTHER_UTILS_HPP_
+#endif  // NAV2_UTIL__SMOOTHER_UTILS_HPP_
